# Much faster picture taking on Android, configurable