--- a/android/src/main/java/org/reactnative/camera/CameraViewManager.java
+++ b/android/src/main/java/org/reactnative/camera/CameraViewManager.java
@@ -142,6 +142,11 @@ public class CameraViewManager extends ViewGroupManager<RNCameraView> {
     view.setUsingCamera2Api(useCamera2Api);
   }
 
+  @ReactProp(name = "camera1ScanMode")
+  public void setCamera1ScanMode(RNCameraView view, String camera1ScanMode) {
+    view.setCamera1ScanMode(camera1ScanMode);
+  }
+
   @ReactProp(name = "playSoundOnCapture")
   public void setPlaySoundOnCapture(RNCameraView view, boolean playSoundOnCapture) {
     view.setPlaySoundOnCapture(playSoundOnCapture);