@Boris-c 2019-09-25
@bkDJ Values we’re obtained on a Nokia 7.1 Android one.