@cristianoccazinsp 2019-09-26
This is very interesting. Is there a lot of overhead for having the preview running?

Also, was the measured time done with Camera1 implementation? Camera1 is quite fast right now, almost instant with the latest changes that removed unnecessary focus attempts.