@Boris-c 2019-09-27
@cristianoccazinsp The overhead here comes from processing the frame preview (YUV to RGB and resizing). The overhead is minimal in eco and fast modes, bigger in boost.

The proposed implementation works with Camera1 only.