@cristianoccazinsp 2019-09-27
@Boris-c default resolution, and I'm even doing post processing with image resizing afterwards with another library.
Test comes from a Pixel 2 and Motorola moto g5. Perhaps the camera on Nokia is heavier by itself, hard to tell, but I'm definitely not getting those high numbers.