@Boris-c 2019-09-27
@cristianoccazinsp I think no overhead can be perceived by the user in eco and fast modes