@Boris-c 2019-09-27
It seems your crash happen in ResolveTakenPictureAsyncTask.java, in the section of the code that handles the option `skipProcessing`. I'm not sure because I don't know what version of that file you have. But if that's the case then try to remove `skipProcessing` from your options.

I developed that improvement last year, before that option was available. skipProcessing relies on receiving a byte array in `mImageData`, whereas my method sends a bitmap in `mBitmap`.