@cristianoccazinsp 2019-09-27
Still getting the same crash with skipProcessing={false}.

I'm using your master branch directly from github, so I should be using exactly your code.