@Boris-c 2019-09-27
<img width="771" alt="image" src="https://user-images.githubusercontent.com/7315069/65786371-cfb9ad00-e156-11e9-9c53-b63aac851e7a.png">

Line 75 (from your crash report above) is in the section that handles skipProcessing. Are you sure you've deactivated skipProcess? Are you sure that it's the same crash?