@Boris-c 2019-09-27
Oh, but I see, they did something wrong, they just check `mOptions.hasKey("skipProcessing")` instead of checking its value.

Try to remove the option instead of setting it to false!...