@Boris-c 2019-09-27
It should be `mOptions.hasKey("skipProcessing") && mOptions.getBoolean("skipProcessing")` or something of the kind.