@cristianoccazinsp 2019-09-27
You are absolutely right, that's actually a good catch/bug.

I now see where you were getting such high times, if skipProcessing is not used, the camera is definitely very slow, but it is quite fast with skipProcessing. It is really not Android's camera API fault's but rather all the processing done by the library itself.

New times, without skipProcessing:

none: 2.4 seconds (against 1.1, holy cow skipProcessing makes a huge difference)
eco: 0.9
fast: 0.7
boost: 0.5

I've noticed, however, picture quality is considerably lower with anything that's not set to "none". It is darker, blurrier, and lower resolution (a few hunder kbs lower)

Original:
![1569603103737](https://user-images.githubusercontent.com/48869228/65787226-d3d1d480-e12e-11e9-89a9-ab7ce20229c9.JPEG)

Fast/Boost:
![1569603397263](https://user-images.githubusercontent.com/48869228/65787245-dcc2a600-e12e-11e9-907a-2cb22e79a75e.JPEG)
![1569603275471](https://user-images.githubusercontent.com/48869228/65787246-dd5b3c80-e12e-11e9-89d5-e2c90174377b.JPEG)

Are you willing to do fixes so it also works with skipProcessing? Have you also tried with back/front/wide lense cameras? Your code is not up to date to master so I can't test with the camera selection options.