@cristianoccazinsp 2019-09-27
There's also another subtlety. When using any mode != 'none', on Android 10, the camera no longer makes the capture sound. I believe Android 10 is now copying iOS and any app that uses use of the camera API is forced to make a sound.

Testing on a Pixel 2 I've also got:

none / no skipProcessing: 1.1 seconds
none / skipProcessing: 1 second
boost / no skipProcessing: 0.25 seconds

Again, the picture quality is severely reduced with eco/fast/boost (mostly light and size).

I've gotta say it's a good trade off if you need very quick photos. The API might get a bit complicated with so many options (skip processing, fix orientation, scan mode).