@Boris-c 2019-09-27
You can try

```
                if (mImageData != null) {
                    // Save byte array (it is already a JPEG)
                    fOut.write(mImageData);
                } else if (mBitmap != null) {
                    // Convert to JPEG and save
                    mBitmap.compress(Bitmap.CompressFormat.JPEG, getQuality(), fOut);
                }
```

... as a replacement of ...

```
                    // Save byte array (it is already a JPEG)
                    fOut.write(mImageData);
```