@Boris-c 2019-09-27
As for the quality of the picture, no idea. I get very good pictures on my device.